test_delete_pet parametrized
